+++ src/lib/utils/clock.cpp
@@ -44,7 +44,7 @@ void clockGettime(struct timespec* t) {
 #endif
 }
 
-long long clockGetMillis() {
+uint64_t clockGetMillis() {
   struct timespec t;
   clockGettime(&t);
   return t.tv_sec*1000LL + t.tv_nsec / 1000000;
